Fix build

+++ db2/core2/fee.go
@@ -7,7 +7,7 @@ type Fee struct {
 	Percent     int64  `db:"percent"`
 	Subtype     int64  `db:"subtype"`
 	AccountID   string `db:"account_id"`
-	AccountType int32  `db:"account_type"`
+	AccountType uint64 `db:"account_type"`
 	LowerBound  int64  `db:"lower_bound"`
 	UpperBound  int64  `db:"upper_bound"`
 	Hash        string `db:"hash"`
